Remove the timers
These are sort of a good idea, but they aren't much effective.
They are noticeable at first, but then you learn to ignore them.
Also the fact that the implementation is super messy.
Also the fact that the actual messages are obfuscated :|
I'd revisit this issue soon, but for now I am removing them.