vim: set shell=/bin/sh
This is displeasing, but it was prompted by a failure when trying the
Recover.vim plugin. It calls `system` from vim with commands which use
the `&&` syntax not supported by the fish shell.

The reason for not using /usr/bin/bash is that it is assumed that most
of the times, the shell is being called by way of `system` and the
use-case served by `system` is sufficed by /bin/sh.

It does mean that running `:sh` now drops into /bin/sh instead of the
user shell, but that's not a huge issue. If it does turn out to be an
inconvenience, we may consider overriding `:sh` itself in future to
automatically exec the user shell.

It's worth noting that the vim wikia explicitly notes this:

> Ensure :echo $SHELL is correct. You will need to use a POSIX shell for
> full Vim compatibility. The fish shell is known to not be compatible.

Ref: https://github.com/chrisbra/Recover.vim/blob/aae97e960b5ede68fff540fb7c162add4be9c973/autoload/recover.vim#L154
Ref: http://vim.wikia.com/wiki/Troubleshooting#vimrc