tmux: use exec in default-command, fix for fish v3 "bug"
Fun bug, took me a while to get it.

After updating to fish v3, opening new terminals in tmux won't open the
shell in the same path as the opener shell. But:

- If I run a nested fish or bash shell inside the opener shell and then
  open a new shell, it does follow the correct path
- Unsetting `default-command` from tmux would also restore the original
  intended behavior
- Same behavior wouldn't occur if default shell is bash or zsh
- Inspecting `${pane_current_path}` would suggest that the variable
  isn't even being updated at all when the shell is fish and
  default-command is set to "$SHELL"!

Turns out, in the original setup, when a new terminal is opened by tmux,
it launches the shell as `fish -c /usr/bin/fish`. So there is the parent
fish shell and then that shell launches `/usr/bin/fish`. Any directory
change done in this nested shell is not tracked by the parent fish shell
instance, which is what tmux would read for `#{pane_current_path}`. This
could be verified still by running `fish -c /usr/bin/fish`, cd'ing in
this new shell, and then opening a new terminal to verify that the new
path hasn't been followed. Running `bash -c /usr/bin/fish` however does
not exhibit this behavior.

Changing `default-command` to `exec $SHELL` eliminates this nesting, and
hence tmux can read the appropriate path. Additionally, this results in
a cleaner process tree as well regardless of which shell is being used.

I am not sure if this is a bug or intended behavior with the new release
of fish. Perhaps it is related to a change[1] in their new handling of
current directory path.

[1]: https://github.com/fish-shell/fish-shell/issues/3350