bash, fish: $INTERACTIVE_SHELL
With this method, on machines where fish does not work very well, we can
simply unset INTERACTIVE_SHELL (or set it to zsh or something).

This change was made because it wasn't feasible to use fish as login
shell on CentOS, and I had to override the tmux and st config to launch
fish.