dmenu-weighted: new script
Wrapper script around dmenu to preserve a weighted history.

Path of history is passed as first argument. Exact command to invoke
dmenu with is the second argument, defaults to "dmenu".

Currently there is no way to seed the input with something. So it just
works like a note taking app by default (even though that is not the
intended use-case).