quickclipsd, quickclips-action: new scripts
quickclips-action uses dmenu to select one of last few clipboard
entries. Immediately after that, another dmenu instance takes a command
and passes the selected clipboard entry to this command. If the command
starts with "| ", the selection is passed as stdin. If the command
consists of the substring "<>", the selection is substituted and the
command is run without any stdin. Else the selection is passed as the
first argument (single token) to the command. In all cases, the command
is executed using `sh -c`.

quickclipsd is the daemon part of it, which listens for clipboard events
using an external program (clipnotify) and stores a configurable number
of clips (defaults to 10), most recent first. The action script reads
this file.

There are several limitations to this daemon, like it doesn't know how
to deal with newlines and dmenu will choke on large amounts of data.
Currently the script doesn't process clipboard data if it is larger than
2000 characters. Ideally clipmenu should be used. However, it just
copies the selected entry from dmenu and doesn't allow for a custom
command to execute.

There are few other issues with clipmenu, adding as reference to this
commit so I can get back to them later. Will fix them and switch to
clipmenu once I get more time, specially since it also supports OpenBSD.

Ref: https://github.com/cdown/clipmenu/issues/109
Ref: https://github.com/cdown/clipmenu/issues/107
Ref: https://github.com/cdown/clipmenu/issues/92